WIP elixir cut: wire emitters deleted, elixirtable.go trimmed; does not build yet (buildVersionModule to recover, sort import unused)
