--- internal/codegen/elixirtable/runtime.go
@@ -1,1614 +1,0 @@
-// The unit's SHARED table runtime (TableRuntime.ex): the read report, the
-// clamps, the skipper, the float bit patterns, and the ONE generic walk over
-// the reflection descriptors that is the text form (docs/SPEC-TABLES.md §16).
-//
-// It is emitted once per UNIT rather than once per file, on the same terms the
-// C# and Rust backends emit theirs: a unit's Elixir modules compile into one
-// application, so a second copy would be a duplicate module rather than C++'s
-// harmless re-inclusion behind a guard. `make tables-elixir-walk` is the gate —
-// the runtime's source, byte-identical across every unit of the corpus, with
-// nothing normalized away but the generated banner.
-package elixirtable
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/mas-bandwidth/schema/v2/ir"
-)
-
-func runtimeModule(u *ir.Unit, ns string) []byte {
-	var b strings.Builder
-	b.WriteString(header(RuntimeModule, u.Package, "the unit's shared table runtime (docs/SPEC-TABLES.md §4, §16)"))
-	b.WriteString("\n")
-	fmt.Fprintf(&b, "defmodule %s.TableRuntime do\n", ns)
-	b.WriteString(runtimeSource)
-	b.WriteString("end\n")
-	return []byte(b.String())
-}
-
-// buildVersionModule carries the unit's BUILD VERSION and nothing else (§20):
-// one digest answering "which build?" and not "which form?", so it belongs to
-// neither accelerator. Both of them compare against it.
-func buildVersionModule(u *ir.Unit, ns string) []byte {
-	var b strings.Builder
-	b.WriteString(header(BuildVersionModule, u.Package, "the unit's build version (docs/SPEC-TABLES.md §20)"))
-	b.WriteString("\n")
-	fmt.Fprintf(&b, "defmodule %s.BuildVersion do\n", ns)
-	b.WriteString("  @moduledoc \"\"\"\n")
-	b.WriteString("  The unit's BUILD VERSION: one digest over every fact a pointed-at form\n")
-	b.WriteString("  depends on (docs/SPEC-TABLES.md §20). A block image or a cooked file whose\n")
-	b.WriteString("  header carries a different one is refused, because its bytes were laid out\n")
-	b.WriteString("  by a build this one does not agree with.\n")
-	b.WriteString("  \"\"\"\n\n")
-	fmt.Fprintf(&b, "  @build_version 0x%016X\n\n", ir.BuildVersion(u))
-	b.WriteString("  def build_version, do: @build_version\n")
-	b.WriteString("end\n")
-	return []byte(b.String())
-}
-
-const runtimeSource = `  @moduledoc """
-  The unit's shared table runtime: the read report (docs/SPEC-TABLES.md §4), the
-  clamps and the skipper the generated codecs call, the IEEE-754 bit patterns
-  the wire carries, and the ONE generic walk over the reflection descriptors
-  that is the text form (§16).
-
-  WHAT ALLOCATES, STATED PLAINLY. The BEAM owns allocation, so the claim a
-  systems language makes — "the read path allocates nothing" — cannot be made
-  here and is not. What IS true, and what the soak gates (make
-  tables-elixir-alloc-audit):
-
-    * a wire READ allocates the decoded value and one SUB-BINARY per framed
-      field. A sub-binary is a small heap term that REFERENCES the caller's
-      bytes; the payload is never copied.
-    * a wire WRITE allocates the binary literals of the framing and one iolist
-      cell per field, flattened once at the surface. No intermediate copy of
-      the payload is made.
-    * the descriptors allocate NOTHING: they are module attributes, so they
-      live in the module's literal area and a read of one copies no term.
-
-  The instrument is the allocation COUNT per iteration, not heap drift: a
-  formatter that allocates and frees the same words every iteration reads zero
-  drift for an hour and is still allocating.
-  """
-
-  import Bitwise
-
-  # ---- the read report (docs/SPEC-TABLES.md §4) ----
-  #
-  # A VALUE, not a member of the reader: the BEAM has no mutable struct, so the
-  # report is threaded down beside the value and handed back. One report, one
-  # caller, the same five counters. duplicate is the TEXT form's counter and a
-  # wire read always leaves it zero (§4).
-  defmodule Report do
-    defstruct unknown: 0, kind_mismatch: 0, clamped: 0, duplicate: 0, malformed: false
-  end
-
-  def report, do: %Report{}
-
-  def unknown(%Report{} = r), do: %Report{r | unknown: r.unknown + 1}
-  def kind_mismatch(%Report{} = r), do: %Report{r | kind_mismatch: r.kind_mismatch + 1}
-  def clamped(%Report{} = r), do: %Report{r | clamped: r.clamped + 1}
-  def duplicate(%Report{} = r), do: %Report{r | duplicate: r.duplicate + 1}
-  def malformed(%Report{} = r), do: %Report{r | malformed: true}
-
-  def silent?(%Report{} = r) do
-    r.unknown == 0 and r.kind_mismatch == 0 and r.clamped == 0 and r.duplicate == 0 and
-      not r.malformed
-  end
-
-  # ---- the clamps (docs/SPEC-TABLES.md §4) ----
-
-  def clamp(value, lo, hi, report) do
-    cond do
-      value < lo -> {lo, clamped(report)}
-      value > hi -> {hi, clamped(report)}
-      true -> {value, report}
-    end
-  end
-
-  def clamp_low(value, lo, report) do
-    if value < lo, do: {lo, clamped(report)}, else: {value, report}
-  end
-
-  def clamp_high(value, hi, report) do
-    if value > hi, do: {hi, clamped(report)}, else: {value, report}
-  end
-
-  # a count past the READER's bound keeps the bounded prefix and counts clamped
-  def clamp_count(count, bound, report) do
-    if count > bound, do: {bound, clamped(report)}, else: {count, report}
-  end
-
-  # a string longer than its declared storage is clamped at the BYTE the wire
-  # gives, which is what the reference does: the wire imposes no encoding (§3)
-  def clamp_bytes(body, bound, report) do
-    if byte_size(body) > bound do
-      {binary_part(body, 0, bound), clamped(report)}
-    else
-      {body, report}
-    end
-  end
-
-  # ---- skipping a field you cannot name (docs/SPEC-TABLES.md §3) ----
-  #
-  # Three rules cover the closed set: kinds 1-11 and 17 skip their fixed width;
-  # 12, 13, 14 and 16 read L and skip L bytes; 15 reads the u16 arm id and stops
-  # there if it is 0, else reads L and skips L. A kind this reader does not know
-  # at all is NOT SKIPPABLE and is framing damage.
-  def skip_payload(kind, data) when kind in [1, 2, 6] do
-    case data do
-      <<_::binary-size(1), rest::binary>> -> rest
-      _ -> :malformed
-    end
-  end
-
-  def skip_payload(kind, data) when kind in [3, 7] do
-    case data do
-      <<_::binary-size(2), rest::binary>> -> rest
-      _ -> :malformed
-    end
-  end
-
-  def skip_payload(kind, data) when kind in [4, 8, 10, 17] do
-    case data do
-      <<_::binary-size(4), rest::binary>> -> rest
-      _ -> :malformed
-    end
-  end
-
-  def skip_payload(kind, data) when kind in [5, 9, 11] do
-    case data do
-      <<_::binary-size(8), rest::binary>> -> rest
-      _ -> :malformed
-    end
-  end
-
-  def skip_payload(kind, data) when kind in [12, 13, 14, 16] do
-    case data do
-      <<len::little-unsigned-32, _::binary-size(len), rest::binary>> -> rest
-      _ -> :malformed
-    end
-  end
-
-  def skip_payload(15, data) do
-    case data do
-      <<0::little-unsigned-16, rest::binary>> ->
-        rest
-
-      <<_arm::little-unsigned-16, len::little-unsigned-32, _::binary-size(len), rest::binary>> ->
-        rest
-
-      _ ->
-        :malformed
-    end
-  end
-
-  def skip_payload(_kind, _data), do: :malformed
-
-  # an id this reader cannot name: skipped by its length, counted
-  def skip_unknown(kind, rest, value, report) do
-    case skip_payload(kind, rest) do
-      :malformed -> {:stop, value, report}
-      rest2 -> {:ok, rest2, value, unknown(report)}
-    end
-  end
-
-  # a field that changed KIND between builds: skipped, never misdecoded, counted
-  def skip_mismatch(kind, rest, value, report) do
-    case skip_payload(kind, rest) do
-      :malformed -> {:stop, value, report}
-      rest2 -> {:ok, rest2, value, kind_mismatch(report)}
-    end
-  end
-
-  # ---- array and keyed bodies ----
-  #
-  # take decodes up to count elements out of one array body. Elements are
-  # BOUNDED by the field body: a count the length cannot cover keeps the decoded
-  # prefix, flags malformed, and the parent continues at the next field, so a
-  # following field's bytes are never fabricated into elements (§4).
-  def take(body, count, report, decode), do: take(body, count, report, decode, [])
-
-  defp take(_body, 0, report, _decode, acc), do: {Enum.reverse(acc), report}
-
-  defp take(body, count, report, decode, acc) do
-    case decode.(body, report) do
-      {:ok, element, tail, report2} ->
-        take(tail, count - 1, report2, decode, [element | acc])
-
-      :short ->
-        {Enum.reverse(acc), malformed(report)}
-    end
-  end
-
-  # keyed walks an enum-keyed body's count pairs of (variant id, L, element).
-  def keyed(_pairs, 0, slots, report, _place), do: {slots, report}
-
-  def keyed(pairs, count, slots, report, place) do
-    case pairs do
-      <<0::little-unsigned-16, _::binary>> ->
-        # None is the NULL KEY: 0 is the reserved id no declared name can fold
-        # to, so a body carrying one is DAMAGED, not merely foreign. Framing
-        # damage stops this body and the parent reads on past the length
-        # (docs/SPEC-TABLES.md §3.2, §4).
-        {slots, malformed(report)}
-
-      <<key::little-unsigned-16, len::little-unsigned-32, element::binary-size(len),
-        tail::binary>> ->
-        {slots2, report2} = place.(key, element, slots, report)
-        keyed(tail, count - 1, slots2, report2, place)
-
-      _ ->
-        {slots, malformed(report)}
-    end
-  end
-
-  def put_slot(slots, index, value), do: List.replace_at(slots, index, value)
-
-  # ---- the IEEE-754 bit patterns the wire carries (docs/SPEC-TABLES.md §3) ----
-  #
-  # A non-finite pattern travels as {:nonfinite, bits}, the convention the
-  # packet emitter established: no BEAM float term can hold one, so the pattern
-  # rides beside the value rather than being lost.
-  def bool_bits(true), do: 1
-  def bool_bits(false), do: 0
-
-  def f32_bits(value) when is_float(value) do
-    <<bits::little-unsigned-32>> = <<value::float-little-32>>
-
-    if (bits >>> 23 &&& 0xFF) == 0xFF do
-      raise ArgumentError, "float overflows float32; write {:nonfinite, bits} instead"
-    end
-
-    bits
-  end
-
-  def f32_bits({:nonfinite, bits}) when bits >= 0 and bits <= 0xFFFFFFFF, do: bits
-
-  def f32_value(bits) do
-    if (bits >>> 23 &&& 0xFF) != 0xFF do
-      <<value::float-little-32>> = <<bits::little-unsigned-32>>
-      value
-    else
-      {:nonfinite, bits}
-    end
-  end
-
-  def f64_bits(value) when is_float(value) do
-    <<bits::little-unsigned-64>> = <<value::float-little-64>>
-    bits
-  end
-
-  def f64_bits({:nonfinite, bits}) when bits >= 0 and bits <= 0xFFFFFFFFFFFFFFFF, do: bits
-
-  def f64_value(bits) do
-    if (bits >>> 52 &&& 0x7FF) != 0x7FF do
-      <<value::float-little-64>> = <<bits::little-unsigned-64>>
-      value
-    else
-      {:nonfinite, bits}
-    end
-  end
-
-  # ---- descriptor access (docs/SPEC-TABLES.md §8) ----
-  #
-  # A descriptor's indirections are {module, function} pairs rather than
-  # captured funs, so the whole structure is a literal a module attribute can
-  # hold: reading one copies no term and allocates nothing.
-  def info({mod, fun}), do: apply(mod, fun, [])
-  def call1({mod, fun}, a), do: apply(mod, fun, [a])
-
-  # ---- the text form (docs/SPEC-TABLES.md §16) ----
-  #
-  # ONE generic walk over the descriptors, for every table in the closure. The
-  # caps the C++ walk carries, mirrored so a text one side refuses is refused by
-  # the other.
-  @max_depth 128
-  @max_key 255
-  @max_number 511
-
-  @base64 ~c"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/"
-
-  # Resolve an input byte without walking the alphabet for each character.
-  @decode64_lookup Enum.reduce(
-                     Enum.with_index(@base64),
-                     Tuple.duplicate(nil, 256),
-                     fn {c, n}, acc -> put_elem(acc, c, n) end
-                   )
-
-  # ---- writing ----
-
-  @doc """
-  Renders one instance as one §16 text: {:ok, binary}, or :error where a value
-  has no text spelling at all — a non-finite float, an enum value or union tag
-  no variant names, a flags bit no variant names — which measure and save
-  refuse for the same reason (§5).
-  """
-  def json_write(value, type) do
-    {:ok, IO.iodata_to_binary([write_value(value, type, 0), ?\n])}
-  catch
-    :refused -> :error
-  end
-
-  defp indent(depth), do: [?\n, List.duplicate("  ", depth)]
-
-  defp write_value(value, type, depth) do
-    if depth > @max_depth, do: throw(:refused)
-    fields = type.fields
-
-    written =
-      Enum.reduce(fields, [], fn f, acc ->
-        cond do
-          f.guard != "" and not guard_holds(value, f.guard) -> acc
-          f.optional and Map.fetch!(value, f.key) == nil -> acc
-          true -> [[write_key(f.json), ": ", write_field(value, f, depth + 1)] | acc]
-        end
-      end)
-
-    case Enum.reverse(written) do
-      [] ->
-        "{}"
-
-      entries ->
-        [
-          ?{,
-          entries
-          |> Enum.map(fn entry -> [indent(depth + 1), entry] end)
-          |> Enum.intersperse(?,),
-          indent(depth),
-          ?}
-        ]
-    end
-  end
-
-  # A GUARD is a back-referenced bool field of the same body, possibly negated
-  # and possibly nested; the descriptor carries the expression as text and this
-  # evaluates it against the instance. Writing elides a field whose guard is
-  # false, exactly as the wire does (§3, §16.2).
-  defp guard_holds(value, expr) do
-    expr
-    |> String.split(" && ")
-    |> Enum.all?(fn term ->
-      case term do
-        "!" <> name -> Map.fetch!(value, String.to_existing_atom(name)) != true
-        name -> Map.fetch!(value, String.to_existing_atom(name)) == true
-      end
-    end)
-  end
-
-  defp write_field(value, f, depth) do
-    slot = Map.fetch!(value, f.key)
-
-    case f.array do
-      :bytes ->
-        write_base64(slot)
-
-      :string ->
-        write_key(slot)
-
-      :keyed ->
-        write_keyed(slot, f, depth)
-
-      :none ->
-        write_scalar(slot, f, depth)
-
-      _ ->
-        elements = if f.tuple, do: Tuple.to_list(slot), else: slot
-
-        case elements do
-          [] ->
-            "[]"
-
-          list ->
-            [
-              ?[,
-              list
-              |> Enum.map(fn element -> [indent(depth + 1), write_scalar(element, f, depth + 1)] end)
-              |> Enum.intersperse(?,),
-              indent(depth),
-              ?]
-            ]
-        end
-    end
-  end
-
-  # An enum-keyed array is an OBJECT keyed by VARIANT NAME (§2.4, §16.2). Every
-  # slot the array KEYS is written, as a fixed array writes every element — the
-  # slots ARE the value. None keys no slot, so nothing is written for it.
-  defp write_keyed(slot, f, depth) do
-    elements = if f.tuple, do: Tuple.to_list(slot), else: slot
-
-    case elements do
-      [] ->
-        "{}"
-
-      list ->
-        [
-          ?{,
-          list
-          |> Enum.with_index()
-          |> Enum.map(fn {element, i} ->
-            name = call1(f.key_enum.name, i + 1)
-            if name == nil, do: throw(:refused)
-            [indent(depth + 1), write_key(name), ": ", write_scalar(element, f, depth + 1)]
-          end)
-          |> Enum.intersperse(?,),
-          indent(depth),
-          ?}
-        ]
-    end
-  end
-
-  defp write_scalar(slot, f, depth) do
-    cond do
-      f.arms != nil ->
-        arms = info(f.arms)
-        tag = Map.fetch!(slot, :type)
-
-        if tag == 0 do
-          "{}"
-        else
-          arm = Enum.at(arms, tag)
-          if arm == nil, do: throw(:refused)
-
-          [
-            ?{,
-            indent(depth + 1),
-            write_key(arm.name),
-            ": ",
-            write_value(Map.fetch!(slot, arm.key), info(arm.table), depth + 1),
-            indent(depth),
-            ?}
-          ]
-        end
-
-      f.table != nil ->
-        write_value(slot, info(f.table), depth)
-
-      f.enum != nil ->
-        # a value no variant names has no text spelling, exactly as it has no
-        # wire identity: the writer REFUSES rather than writing None over it
-        name = call1(f.enum.name, slot)
-        if name == nil, do: throw(:refused)
-        write_key(name)
-
-      f.flags != nil ->
-        write_flags(slot, f, depth)
-
-      f.kind == 1 ->
-        if slot, do: "true", else: "false"
-
-      f.kind == 10 ->
-        write_float(slot, true)
-
-      f.kind == 11 ->
-        write_float(slot, false)
-
-      true ->
-        Integer.to_string(slot)
-    end
-  end
-
-  defp write_flags(bits, f, depth) do
-    if bits == 0 do
-      "[]"
-    else
-      names =
-        Enum.reduce(63..0//-1, [], fn bit, acc ->
-          if (bits >>> bit &&& 1) == 0 do
-            acc
-          else
-            name = Enum.at(f.flags, bit)
-            if name == nil, do: throw(:refused)
-            [[indent(depth + 1), write_key(name)] | acc]
-          end
-        end)
-
-      [?[, Enum.intersperse(names, ?,), indent(depth), ?]]
-    end
-  end
-
-  # A JSON text MUST be valid UTF-8 (RFC 8259 §8.1). The read path is
-  # byte-transparent — the wire imposes no encoding (§3) — so the WRITER is
-  # where that obligation is met: a byte that is not part of a well-formed
-  # sequence is written as U+FFFD, one per bad byte, and never raw (§16.3).
-  def write_key(text) when is_binary(text), do: [?", escape(text, []), ?"]
-  def write_key(text), do: write_key(to_string(text))
-
-  defp escape(<<>>, acc), do: Enum.reverse(acc)
-  defp escape(<<?", rest::binary>>, acc), do: escape(rest, ["\\\"" | acc])
-  defp escape(<<?\\, rest::binary>>, acc), do: escape(rest, ["\\\\" | acc])
-  defp escape(<<0x08, rest::binary>>, acc), do: escape(rest, ["\\b" | acc])
-  defp escape(<<0x0C, rest::binary>>, acc), do: escape(rest, ["\\f" | acc])
-  defp escape(<<?\n, rest::binary>>, acc), do: escape(rest, ["\\n" | acc])
-  defp escape(<<?\r, rest::binary>>, acc), do: escape(rest, ["\\r" | acc])
-  defp escape(<<?\t, rest::binary>>, acc), do: escape(rest, ["\\t" | acc])
-
-  defp escape(<<c, rest::binary>>, acc) when c < 0x20 do
-    escape(rest, [["\\u00", hex_digit(c >>> 4), hex_digit(c &&& 0xF)] | acc])
-  end
-
-  defp escape(<<c, rest::binary>>, acc) when c < 0x80, do: escape(rest, [c | acc])
-
-  defp escape(data, acc) do
-    case utf8_sequence(data) do
-      0 ->
-        <<_bad, rest::binary>> = data
-        escape(rest, [<<0xEF, 0xBF, 0xBD>> | acc])
-
-      width ->
-        <<unit::binary-size(^width), rest::binary>> = data
-        escape(rest, [unit | acc])
-    end
-  end
-
-  defp hex_digit(v) when v < 10, do: ?0 + v
-  defp hex_digit(v), do: ?a + v - 10
-
-  # the length of ONE well-formed UTF-8 sequence at the head, or 0 when the
-  # bytes there are not one: a stray continuation, an overlong form, a surrogate
-  # half, and anything past U+10FFFF are all rejected
-  defp utf8_sequence(<<c::utf8, _::binary>> = data) do
-    width = byte_size(<<c::utf8>>)
-    <<head::binary-size(^width), _::binary>> = data
-    if head == <<c::utf8>>, do: width, else: 0
-  end
-
-  defp utf8_sequence(_), do: 0
-
-  defp write_base64(data), do: [?", Base.encode64(data), ?"]
-
-  # A float writes at the SHORTEST precision that reads back as the same value
-  # at the field's own width, so a round trip is exact and a text stays
-  # readable. Non-finite values have no JSON spelling at all and the writer
-  # REFUSES rather than losing one silently (§16.3).
-  defp write_float({:nonfinite, _bits}, _single), do: throw(:refused)
-
-  defp write_float(value, single) when is_integer(value), do: write_float(value * 1.0, single)
-
-  defp write_float(value, single) do
-    {low, high} = if single, do: {6, 9}, else: {15, 17}
-    shortest(value, single, low, high)
-  end
-
-  defp shortest(value, single, digits, high) do
-    text = format_g(value, digits)
-
-    cond do
-      digits >= high -> text
-      reads_back?(text, value, single) -> text
-      true -> shortest(value, single, digits + 1, high)
-    end
-  end
-
-  defp reads_back?(text, value, single) do
-    case Float.parse(text) do
-      {back, ""} -> if single, do: narrow32(back) == value, else: back == value
-      _ -> false
-    end
-  end
-
-  # one emulated float32 step: a double that no float32 can hold rounds to the
-  # infinity pattern on the way in, and NO BEAM FLOAT TERM CAN HOLD ONE, so the
-  # answer is the atom rather than a value. The text form's read turns that into
-  # a kind mismatch, which is what keeps an infinity out of storage (§16.2).
-  defp narrow32(v) do
-    <<bits::little-unsigned-32>> = <<v::float-little-32>>
-
-    if (bits >>> 23 &&& 0xFF) == 0xFF do
-      :infinity
-    else
-      <<narrow::float-little-32>> = <<bits::little-unsigned-32>>
-      narrow
-    end
-  rescue
-    ArgumentError -> :infinity
-  end
-
-  # C's "%.*g", exactly, because the goldens are its bytes: convert to
-  # precision significant digits, choose style e when the decimal exponent is
-  # below -4 or at least the precision and style f otherwise, then strip the
-  # fraction's trailing zeros and a bare decimal point. The exponent field
-  # carries a sign and at least two digits.
-  defp format_g(value, precision) do
-    precision = max(precision, 1)
-    scientific = :erlang.float_to_binary(value, scientific: precision - 1)
-    [mantissa, exponent_text] = String.split(scientific, "e")
-    exponent = String.to_integer(exponent_text)
-
-    if exponent < -4 or exponent >= precision do
-      trim_zeros(mantissa) <> "e" <> exponent_digits(exponent)
-    else
-      trim_zeros(:erlang.float_to_binary(value, decimals: max(precision - 1 - exponent, 0)))
-    end
-  end
-
-  defp trim_zeros(text) do
-    if String.contains?(text, ".") do
-      trimmed = String.replace_trailing(text, "0", "")
-      if String.ends_with?(trimmed, "."), do: binary_part(trimmed, 0, byte_size(trimmed) - 1), else: trimmed
-    else
-      text
-    end
-  end
-
-  defp exponent_digits(exponent) do
-    sign = if exponent < 0, do: "-", else: "+"
-    magnitude = abs(exponent)
-    digits = Integer.to_string(magnitude)
-    if magnitude < 10, do: sign <> "0" <> digits, else: sign <> digits
-  end
-
-  # ---- reading ----
-
-  @doc """
-  Fills one instance from one §16 text. Returns {value, report}; report.malformed
-  says the text is not JSON, and the instance holds what was placed before the
-  stop. Fields the text does not mention keep their declared defaults, exactly
-  as an absent field on the wire does (§4).
-  """
-  def json_read(text, type) do
-    value = info(type.new)
-    report = report()
-
-    try do
-      {value2, pos, report2} = read_table(value, type, text, 0, 0, report)
-      pos2 = space(text, pos)
-
-      if pos2 != byte_size(text) do
-        # trailing rubbish is not one text
-        {value2, malformed(report2)}
-      else
-        {value2, report2}
-      end
-    catch
-      {:bad, value2, report2} -> {value2, malformed(report2)}
-    end
-  end
-
-  defp bad(value, report), do: throw({:bad, value, report})
-
-  defp at(text, pos) do
-    if pos < byte_size(text), do: :binary.at(text, pos), else: 0
-  end
-
-  defp space(text, pos) do
-    case at(text, pos) do
-      c when c in [?\s, ?\t, ?\n, ?\r] -> space(text, pos + 1)
-      _ -> pos
-    end
-  end
-
-  defp peek(text, pos) do
-    pos = space(text, pos)
-    {at(text, pos), pos}
-  end
-
-  # classify the value at the cursor without consuming it
-  defp value_shape(text, pos) do
-    {c, pos} = peek(text, pos)
-
-    shape =
-      case c do
-        ?{ -> :object
-        ?[ -> :array
-        ?" -> :string
-        ?t -> :bool
-        ?f -> :bool
-        ?n -> :null
-        0 -> :end
-        _ -> :number
-      end
-
-    {shape, pos}
-  end
-
-  defp literal(text, pos, word, value, report) do
-    size = byte_size(word)
-
-    if pos + size <= byte_size(text) and binary_part(text, pos, size) == word do
-      pos + size
-    else
-      bad(value, report)
-    end
-  end
-
-  # ---- one JSON string, one code point at a time ----
-  #
-  # A string longer than the field is clamped AT A CODE POINT BOUNDARY and never
-  # cut through a multi-byte character, which is why the scan appends whole
-  # units rather than bytes. capacity < 0 is unbounded.
-  defp scan_string(text, pos, capacity, value, report) do
-    {c, pos} = peek(text, pos)
-    if c != ?", do: bad(value, report)
-    scan_body(text, pos + 1, capacity, [], 0, false, value, report)
-  end
-
-  defp scan_body(text, pos, capacity, acc, used, clamped, value, report) do
-    case at(text, pos) do
-      0 when pos >= byte_size(text) ->
-        bad(value, report)
-
-      ?" ->
-        {IO.iodata_to_binary(Enum.reverse(acc)), pos + 1, clamped}
-
-      ?\\ ->
-        {unit, pos2} = scan_escape(text, pos + 1, value, report)
-        keep(text, pos2, capacity, acc, used, clamped, unit, value, report)
-
-      c when c < 0x20 ->
-        # a raw control character is not a JSON string body
-        bad(value, report)
-
-      c ->
-        want =
-          cond do
-            (c &&& 0xE0) == 0xC0 -> 2
-            (c &&& 0xF0) == 0xE0 -> 3
-            (c &&& 0xF8) == 0xF0 -> 4
-            true -> 1
-          end
-
-        {unit, pos2} = take_unit(text, pos + 1, want - 1, <<c>>)
-        keep(text, pos2, capacity, acc, used, clamped, unit, value, report)
-    end
-  end
-
-  # only bytes that ACTUALLY look like continuations are taken: the wire imposes
-  # no encoding, so a string may legitimately hold a stray lead byte, and one at
-  # the end of a text must not swallow the closing quote
-  defp take_unit(_text, pos, 0, unit), do: {unit, pos}
-
-  defp take_unit(text, pos, left, unit) do
-    c = at(text, pos)
-
-    if pos < byte_size(text) and (c &&& 0xC0) == 0x80 do
-      take_unit(text, pos + 1, left - 1, unit <> <<c>>)
-    else
-      {unit, pos}
-    end
-  end
-
-  defp keep(text, pos, capacity, acc, used, clamped, unit, value, report) do
-    size = byte_size(unit)
-
-    if capacity >= 0 and used + size > capacity do
-      scan_body(text, pos, capacity, acc, used, true, value, report)
-    else
-      scan_body(text, pos, capacity, [unit | acc], used + size, clamped, value, report)
-    end
-  end
-
-  defp scan_escape(text, pos, value, report) do
-    case at(text, pos) do
-      ?" -> {"\"", pos + 1}
-      ?\\ -> {"\\", pos + 1}
-      ?/ -> {"/", pos + 1}
-      ?b -> {<<0x08>>, pos + 1}
-      ?f -> {<<0x0C>>, pos + 1}
-      ?n -> {"\n", pos + 1}
-      ?r -> {"\r", pos + 1}
-      ?t -> {"\t", pos + 1}
-      ?u -> scan_unicode(text, pos + 1, value, report)
-      _ -> bad(value, report)
-    end
-  end
-
-  defp scan_unicode(text, pos, value, report) do
-    high = hex4(text, pos, value, report)
-    pos = pos + 4
-
-    {code, pos} =
-      if high >= 0xD800 and high <= 0xDBFF and at(text, pos) == ?\\ and at(text, pos + 1) == ?u do
-        low = hex4(text, pos + 2, value, report)
-
-        if low >= 0xDC00 and low <= 0xDFFF do
-          {0x10000 + ((high - 0xD800) <<< 10) + (low - 0xDC00), pos + 6}
-        else
-          # a lone lead surrogate never found its partner
-          {high, pos}
-        end
-      else
-        {high, pos}
-      end
-
-    # a surrogate half with no partner has no UTF-8 encoding: encoding it anyway
-    # would manufacture CESU-8 out of input that was valid JSON, so it reads as
-    # the replacement character (§16.3)
-    code = if code >= 0xD800 and code <= 0xDFFF, do: 0xFFFD, else: code
-    {<<code::utf8>>, pos}
-  end
-
-  defp hex4(text, pos, value, report) do
-    if pos + 4 > byte_size(text), do: bad(value, report)
-
-    Enum.reduce(0..3, 0, fn i, acc ->
-      c = :binary.at(text, pos + i)
-
-      digit =
-        cond do
-          c >= ?0 and c <= ?9 -> c - ?0
-          c >= ?a and c <= ?f -> c - ?a + 10
-          c >= ?A and c <= ?F -> c - ?A + 10
-          true -> bad(value, report)
-        end
-
-      (acc <<< 4) + digit
-    end)
-  end
-
-  # ---- one RFC 8259 number ----
-  #
-  # Scanning the production is what makes a typo in an authoring file a
-  # DIAGNOSTIC rather than a value: "1-2" scans as 1 and leaves "-2" where the
-  # object expects a comma, so the text is malformed (§16.2). A permissive scan
-  # would hand "1-2" to a digit loop and report a clamp.
-  defp walk_number(text, pos) do
-    pos = space(text, pos)
-    pos = if at(text, pos) == ?-, do: pos + 1, else: pos
-
-    case at(text, pos) do
-      ?0 ->
-        walk_frac(text, pos + 1, true)
-
-      c when c >= ?1 and c <= ?9 ->
-        walk_frac(text, digits(text, pos), true)
-
-      _ ->
-        :bad
-    end
-  end
-
-  defp digits(text, pos) do
-    c = at(text, pos)
-    if c >= ?0 and c <= ?9, do: digits(text, pos + 1), else: pos
-  end
-
-  defp walk_frac(text, pos, integral) do
-    if at(text, pos) == ?. do
-      after_point = digits(text, pos + 1)
-      if after_point == pos + 1, do: :bad, else: walk_exp(text, after_point, false)
-    else
-      walk_exp(text, pos, integral)
-    end
-  end
-
-  defp walk_exp(text, pos, integral) do
-    c = at(text, pos)
-
-    if c == ?e or c == ?E do
-      pos2 = pos + 1
-      pos2 = if at(text, pos2) in [?-, ?+], do: pos2 + 1, else: pos2
-      after_digits = digits(text, pos2)
-      if after_digits == pos2, do: :bad, else: {after_digits, false}
-    else
-      {pos, integral}
-    end
-  end
-
-  defp scan_number(text, pos, value, report) do
-    start = space(text, pos)
-
-    case walk_number(text, start) do
-      :bad ->
-        bad(value, report)
-
-      {stop, integral} ->
-        size = stop - start
-        if size <= 0 or size > @max_number, do: bad(value, report)
-        {binary_part(text, start, size), integral, stop}
-    end
-  end
-
-  # ---- skipping a value the walk will not place ----
-
-  defp skip_value(text, pos, depth, value, report) do
-    {c, pos} = peek(text, pos)
-
-    case c do
-      ?{ -> skip_container(text, pos + 1, ?}, depth, value, report)
-      ?[ -> skip_container(text, pos + 1, ?], depth, value, report)
-      ?" -> elem(scan_string(text, pos, 0, value, report), 1)
-      ?t -> literal(text, pos, "true", value, report)
-      ?f -> literal(text, pos, "false", value, report)
-      ?n -> literal(text, pos, "null", value, report)
-      0 -> bad(value, report)
-      _ -> elem(scan_number(text, pos, value, report), 2)
-    end
-  end
-
-  defp skip_container(text, pos, close, depth, value, report) do
-    if depth > @max_depth, do: bad(value, report)
-    {c, pos} = peek(text, pos)
-
-    cond do
-      c == close ->
-        pos + 1
-
-      c == 0 ->
-        bad(value, report)
-
-      true ->
-        pos =
-          if close == ?} do
-            {_key, pos2, _} = scan_string(text, pos, 0, value, report)
-            {colon, pos3} = peek(text, pos2)
-            if colon != ?:, do: bad(value, report)
-            pos3 + 1
-          else
-            pos
-          end
-
-        pos = skip_value(text, pos, depth + 1, value, report)
-        {c2, pos2} = peek(text, pos)
-
-        cond do
-          # a trailing comma is accepted
-          c2 == ?, -> skip_container(text, pos2 + 1, close, depth, value, report)
-          c2 == close -> pos2 + 1
-          true -> bad(value, report)
-        end
-    end
-  end
-
-  # ---- one table object ----
-
-  defp read_table(value, type, text, pos, depth, report) do
-    if depth > @max_depth, do: bad(value, report)
-    {c, pos} = peek(text, pos)
-    if c != ?{, do: bad(value, report)
-    read_entries(value, type, text, pos + 1, depth, report, %{})
-  end
-
-  defp read_entries(value, type, text, pos, depth, report, seen) do
-    {c, pos} = peek(text, pos)
-
-    cond do
-      c == ?} ->
-        {value, pos + 1, report}
-
-      c == 0 ->
-        bad(value, report)
-
-      true ->
-        {key, pos, _} = scan_string(text, pos, @max_key, value, report)
-        {colon, pos} = peek(text, pos)
-        if colon != ?:, do: bad(value, report)
-        pos = pos + 1
-        # THE AMPERSAND PREFIX IS RESERVED TO THE FORM (docs/SPEC-TABLES.md
-        # 16.7): never a field this build lacks, always a construct it cannot
-        # honor. Malformed and refused, never counted as unknown.
-        # Two guard BIFs and not a binary match: matching the key would put a
-        # match context on the heap per key, and the read path's heap words are
-        # pinned (test/conformance/elixir/alloc-budget.txt).
-        if byte_size(key) > 0 and :binary.first(key) == ?&, do: bad(value, report)
-
-        {value, pos, report, seen} =
-          case Enum.find(type.fields, fn f -> f.json == key end) do
-            nil ->
-              # a key this reader cannot name
-              {value, skip_value(text, pos, depth + 1, value, report), unknown(report), seen}
-
-            f ->
-              report = if Map.has_key?(seen, f.key), do: duplicate(report), else: report
-              seen = Map.put(seen, f.key, true)
-              {value, pos, report} = read_entry(value, f, text, pos, depth, report)
-              {value, pos, report, seen}
-          end
-
-        {c2, pos2} = peek(text, pos)
-
-        cond do
-          c2 == ?, -> read_entries(value, type, text, pos2 + 1, depth, report, seen)
-          c2 == ?} -> {value, pos2 + 1, report}
-          true -> bad(value, report)
-        end
-    end
-  end
-
-  defp read_entry(value, f, text, pos, depth, report) do
-    {shape, pos} = value_shape(text, pos)
-
-    cond do
-      shape == :null and f.optional ->
-        # null is the absence, not a value: a ?T reads it as ABSENT (§16.2). It
-        # is the ONE key that puts a field back at its defaults, so a repeated
-        # key whose last occurrence is null cannot leave an earlier value.
-        {Map.put(value, f.key, nil), literal(text, pos, "null", value, report), report}
-
-      shape != field_shape(f) ->
-        # the wrong JSON type for the kind: skipped, never coerced
-        {value, skip_value(text, pos, depth + 1, value, report), kind_mismatch(report)}
-
-      true ->
-        read_field(value, f, text, pos, depth, report)
-    end
-  end
-
-  # the JSON shape a field's declaration takes. A ?T takes T's shape — presence
-  # is the KEY's, never the value's — and an enum-keyed array is an OBJECT.
-  defp field_shape(f) do
-    cond do
-      f.array == :string or f.array == :bytes -> :string
-      f.array == :keyed -> :object
-      f.array != :none -> :array
-      f.arms != nil -> :object
-      f.table != nil -> :object
-      f.enum != nil -> :string
-      f.flags != nil -> :array
-      f.kind == 1 -> :bool
-      true -> :number
-    end
-  end
-
-  # the same classifier one level down, for an array's elements
-  defp element_shape(f) do
-    cond do
-      f.arms != nil -> :object
-      f.table != nil -> :object
-      f.enum != nil -> :string
-      f.flags != nil -> :array
-      f.kind == 1 -> :bool
-      true -> :number
-    end
-  end
-
-  defp read_field(value, f, text, pos, depth, report) do
-    case f.array do
-      :string ->
-        {out, pos, clamped} = scan_string(text, pos, f.bound, value, report)
-        report = if clamped, do: clamped(report), else: report
-        {Map.put(value, f.key, out), pos, report}
-
-      :bytes ->
-        read_base64(value, f, text, pos, report)
-
-      :keyed ->
-        read_keyed(value, f, text, pos, depth, report)
-
-      :none ->
-        {slot, pos, report} = read_scalar(Map.fetch!(value, f.key), f, text, pos, depth, value, report)
-        {Map.put(value, f.key, slot), pos, report}
-
-      _ ->
-        read_array(value, f, text, pos, depth, report)
-    end
-  end
-
-  # a bytes(N) body decoded six bits at a time, exactly as the generated walk
-  # does: a body that is not base64 is the wrong SHAPE for the kind, so the
-  # field keeps its default and the event is counted
-  defp read_base64(value, f, text, pos, report) do
-    {c, pos} = peek(text, pos)
-    if c != ?", do: bad(value, report)
-
-    case validate_base64(text, pos + 1, 0, 0, false, value, report) do
-      {:ok, end_pos} ->
-        decode64(value, f, text, pos + 1, end_pos, 0, 0, [], 0, false, report)
-
-      {:mismatch, end_pos} ->
-        {value, end_pos, kind_mismatch(report)}
-    end
-  end
-
-  defp validate_base64(text, pos, symbols, pad, malformed, value, report) do
-    if pos >= byte_size(text), do: bad(value, report)
-
-    case :binary.at(text, pos) do
-      ?" ->
-        cond do
-          malformed ->
-            {:mismatch, pos + 1}
-
-          pad > 0 and (rem(symbols + pad, 4) != 0 or rem(symbols, 4) < 2 or pad > 2) ->
-            {:mismatch, pos + 1}
-
-          pad == 0 and rem(symbols, 4) == 1 ->
-            {:mismatch, pos + 1}
-
-          true ->
-            {:ok, pos + 1}
-        end
-
-      ?= ->
-        validate_base64(text, pos + 1, symbols, pad + 1, malformed, value, report)
-
-      c ->
-        if pad > 0 do
-          validate_base64(text, pos + 1, symbols, pad, true, value, report)
-        else
-          case elem(@decode64_lookup, c) do
-            nil ->
-              validate_base64(text, pos + 1, symbols, 0, true, value, report)
-
-            _at ->
-              validate_base64(text, pos + 1, symbols + 1, 0, malformed, value, report)
-          end
-        end
-    end
-  end
-
-  defp decode64(value, f, text, pos, end_pos, acc, held, out, used, clamped, report) do
-    c = :binary.at(text, pos)
-
-    cond do
-      c == ?" or c == ?= ->
-        report = if clamped, do: clamped(report), else: report
-        {Map.put(value, f.key, IO.iodata_to_binary(Enum.reverse(out))), end_pos, report}
-
-      true ->
-        at = elem(@decode64_lookup, c)
-        acc2 = (acc <<< 6) ||| at
-        held2 = held + 6
-
-        if held2 >= 8 do
-          held3 = held2 - 8
-          byte = acc2 >>> held3 &&& 0xFF
-          acc2 = acc2 &&& ((1 <<< held3) - 1)
-
-          if used < f.bound do
-            decode64(value, f, text, pos + 1, end_pos, acc2, held3, [byte | out], used + 1, clamped, report)
-          else
-            decode64(value, f, text, pos + 1, end_pos, acc2, held3, out, used, true, report)
-          end
-        else
-          decode64(value, f, text, pos + 1, end_pos, acc2, held2, out, used, clamped, report)
-        end
-    end
-  end
-
-  # an ARRAY re-establishes before placing: a fixed array writes every slot, so
-  # a second, shorter occurrence overlaying a prefix would leave the first
-  # occurrence's tail standing (§16.2, last-wins as a WHOLE value)
-  defp read_array(value, f, text, pos, depth, report) do
-    {c, pos} = peek(text, pos)
-    if c != ?[, do: bad(value, report)
-    {placed, pos, report} = read_elements(value, f, text, pos + 1, depth, report, [], 0)
-    list = Enum.reverse(placed)
-
-    filled =
-      if f.array == :fixed do
-        # a fixed array's tail keeps the declared defaults, exactly as a short
-        # wire count does
-        list ++ List.duplicate(f.elem_default, f.bound - length(list))
-      else
-        list
-      end
-
-    {Map.put(value, f.key, filled), pos, report}
-  end
-
-  defp read_elements(value, f, text, pos, depth, report, acc, placed) do
-    {c, pos} = peek(text, pos)
-
-    cond do
-      c == ?] ->
-        {acc, pos + 1, report}
-
-      c == 0 ->
-        bad(value, report)
-
-      true ->
-        {acc, pos, report, placed} =
-          cond do
-            placed >= f.bound ->
-              # more elements than the reader's bound: the bounded prefix is
-              # kept and the excess counts, the wire's rule (§4)
-              {acc, skip_value(text, pos, depth + 1, value, report), clamped(report), placed}
-
-            elem(value_shape(text, pos), 0) != element_shape(f) ->
-              {[f.elem_default | acc], skip_value(text, pos, depth + 1, value, report),
-               kind_mismatch(report), placed + 1}
-
-            true ->
-              {slot, pos2, report2} =
-                read_scalar(f.elem_default, f, text, pos, depth + 1, value, report)
-
-              {[slot | acc], pos2, report2, placed + 1}
-          end
-
-        {c2, pos2} = peek(text, pos)
-
-        cond do
-          c2 == ?, -> read_elements(value, f, text, pos2 + 1, depth, report, acc, placed)
-          c2 == ?] -> {acc, pos2 + 1, report}
-          true -> bad(value, report)
-        end
-    end
-  end
-
-  # an enum-keyed array is an OBJECT keyed by VARIANT NAME. An absent key keeps
-  # that slot's defaults; an unknown key is skipped and counted, and "None" is
-  # such a key because None keys no slot (§2.4).
-  defp read_keyed(value, f, text, pos, depth, report) do
-    {c, pos} = peek(text, pos)
-    if c != ?{, do: bad(value, report)
-    slots = List.duplicate(f.elem_default, f.bound)
-    {slots, pos, report} = read_slots(value, f, text, pos + 1, depth, report, slots, %{})
-    {Map.put(value, f.key, if(f.tuple, do: List.to_tuple(slots), else: slots)), pos, report}
-  end
-
-  defp read_slots(value, f, text, pos, depth, report, slots, seen) do
-    {c, pos} = peek(text, pos)
-
-    cond do
-      c == ?} ->
-        {slots, pos + 1, report}
-
-      c == 0 ->
-        bad(value, report)
-
-      true ->
-        {key, pos, _} = scan_string(text, pos, @max_key, value, report)
-        {colon, pos} = peek(text, pos)
-        if colon != ?:, do: bad(value, report)
-        pos = pos + 1
-        variant = call1(f.key_enum.value, key)
-        index = if variant == nil or variant < 1 or variant > f.bound, do: nil, else: variant - 1
-
-        report =
-          if index != nil and Map.has_key?(seen, index), do: duplicate(report), else: report
-
-        seen = if index == nil, do: seen, else: Map.put(seen, index, true)
-
-        {slots, pos, report} =
-          cond do
-            index == nil ->
-              # a key this enum cannot name; "None" is one of them
-              {slots, skip_value(text, pos, depth + 1, value, report), unknown(report)}
-
-            elem(value_shape(text, pos), 0) != element_shape(f) ->
-              {slots, skip_value(text, pos, depth + 1, value, report), kind_mismatch(report)}
-
-            true ->
-              {slot, pos2, report2} =
-                read_scalar(Enum.at(slots, index), f, text, pos, depth + 1, value, report)
-
-              {List.replace_at(slots, index, slot), pos2, report2}
-          end
-
-        {c2, pos2} = peek(text, pos)
-
-        cond do
-          c2 == ?, -> read_slots(value, f, text, pos2 + 1, depth, report, slots, seen)
-          c2 == ?} -> {slots, pos2 + 1, report}
-          true -> bad(value, report)
-        end
-    end
-  end
-
-  defp read_scalar(slot, f, text, pos, depth, value, report) do
-    cond do
-      f.arms != nil ->
-        read_union(slot, f, text, pos, depth, value, report)
-
-      f.table != nil ->
-        type = info(f.table)
-        nested = info(type.new)
-        {nested, pos, report} = read_table(nested, type, text, pos, depth + 1, report)
-        {nested, pos, report}
-
-      f.enum != nil ->
-        {name, pos, _} = scan_string(text, pos, @max_key, value, report)
-
-        case call1(f.enum.value, name) do
-          nil ->
-            # a name this build cannot name reads as None and counts as unknown,
-            # exactly as an unknown variant id does on the wire (§4)
-            {0, pos, unknown(report)}
-
-          v ->
-            {v, pos, report}
-        end
-
-      f.flags != nil ->
-        read_flags(f, text, pos, depth, value, report)
-
-      f.kind == 1 ->
-        {c, pos} = peek(text, pos)
-
-        if c == ?t do
-          {true, literal(text, pos, "true", value, report), report}
-        else
-          {false, literal(text, pos, "false", value, report), report}
-        end
-
-      true ->
-        {token, integral, pos} = scan_number(text, pos, value, report)
-
-        cond do
-          f.kind == 10 -> {place_float(token, f, true, report), pos, nil} |> settle()
-          f.kind == 11 -> {place_float(token, f, false, report), pos, nil} |> settle()
-          true -> {place_integer(token, integral, f, slot, report), pos, nil} |> settle()
-        end
-    end
-  end
-
-  defp settle({{slot, report}, pos, nil}), do: {slot, pos, report}
-
-  # a magnitude the field's format cannot hold is the WRONG SHAPE for the kind
-  # and never reaches storage: storing the infinity the conversion produced
-  # would leave an instance this walk called CLEAN that the writer then refuses
-  # forever (§16.2, §16.3)
-  defp place_float(token, f, single, report) do
-    case decimal_to_float(token, single) do
-      :infinity ->
-        {0.0, kind_mismatch(report)}
-
-      parsed ->
-        clamp_range(parsed, f, report)
-    end
-  end
-
-  # ---- decimal -> binary float, ONE correctly-rounded conversion ----
-  #
-  # At the field's OWN width, over exact integers. Converting through a double
-  # first would round TWICE, and the two roundings part company across a whole
-  # band: a decimal between FLT_MAX and the float32 rounding midpoint lands ON
-  # the midpoint as a double and overflows from there, and again at the
-  # subnormal end, where the double rounds to zero and the single does not.
-  # The page's rule is exact conversion at both float widths (§16.2), and the
-  # BEAM offers no strtof, so this is it.
-  def decimal_to_float(token, single) do
-    {precision, min_ulp, overflow_exp} =
-      if single, do: {24, -149, 128}, else: {53, -1074, 1024}
-
-    {sign, m, e} = decimal_parts(token)
-
-    if m == 0 do
-      0.0 * sign
-    else
-      {n, d} = if e >= 0, do: {m * pow10(e), 1}, else: {m, pow10(-e)}
-      ulp = max(exponent_of(n, d) - (precision - 1), min_ulp)
-      {q, ulp} = normalize(round_half_even(n, d, ulp), ulp, precision)
-
-      # q * 2^ulp >= 2^overflow_exp has no float of this width: the value
-      # overflows, and no BEAM float term can hold the infinity
-      room = overflow_exp - ulp
-
-      if room <= 0 or q >= 1 <<< room do
-        :infinity
-      else
-        sign * q * :math.pow(2.0, ulp)
-      end
-    end
-  end
-
-  defp decimal_parts(token) do
-    {sign, rest} =
-      case token do
-        "-" <> r -> {-1, r}
-        "+" <> r -> {1, r}
-        r -> {1, r}
-      end
-
-    {mantissa, exponent} =
-      case String.split(rest, ["e", "E"]) do
-        [m] -> {m, "0"}
-        [m, x] -> {m, x}
-      end
-
-    {whole, fraction} =
-      case String.split(mantissa, ".") do
-        [i] -> {i, ""}
-        [i, f] -> {i, f}
-      end
-
-    digits = whole <> fraction
-    m = if digits == "", do: 0, else: String.to_integer(digits)
-    {sign, m, String.to_integer(exponent) - byte_size(fraction)}
-  end
-
-  defp pow10(k), do: Integer.pow(10, k)
-
-  # n/d >= 2^k, over integers
-  defp ge_pow2(n, d, k) do
-    if k >= 0, do: n >= d <<< k, else: n <<< -k >= d
-  end
-
-  # the k with 2^k <= n/d < 2^(k+1)
-  defp exponent_of(n, d) do
-    k = bit_length(n) - bit_length(d)
-    if ge_pow2(n, d, k), do: k, else: k - 1
-  end
-
-  defp bit_length(0), do: 0
-  defp bit_length(x), do: byte_size(Integer.to_string(x, 2))
-
-  # round(n / (d * 2^ulp)) to nearest, ties to even
-  defp round_half_even(n, d, ulp) do
-    {num, den} = if ulp >= 0, do: {n, d <<< ulp}, else: {n <<< -ulp, d}
-    q = div(num, den)
-    r = rem(num, den)
-    twice = r * 2
-
-    cond do
-      twice > den -> q + 1
-      twice < den -> q
-      rem(q, 2) == 0 -> q
-      true -> q + 1
-    end
-  end
-
-  # a carry out of the top bit widens the exponent by one
-  defp normalize(q, ulp, precision) do
-    if q >= 1 <<< precision, do: {div(q, 2), ulp + 1}, else: {q, ulp}
-  end
-
-  defp clamp_range(value, f, report) do
-    case f.range do
-      nil -> {value, report}
-      {lo, hi} -> clamp(value, lo, hi, report)
-    end
-  end
-
-  # JSON HAS ONE NUMBER TYPE, so an integer field takes any token whose VALUE is
-  # integral, however it was spelled — 2, 2.0 and 1e3 are all the integers this
-  # walk places — and only a genuinely fractional value is the wrong shape for
-  # it (§16.2)
-  defp place_integer(token, integral, f, slot, report) do
-    {value, saturated, ok} =
-      if integral do
-        token_integer(token, f)
-      else
-        case Float.parse(token) do
-          {d, ""} ->
-            if d == Float.round(d) and abs(d) < 1.0e18 do
-              {trunc(d), false, true}
-            else
-              if abs(d) >= 1.0e18, do: {saturate(d, f), true, true}, else: {0, false, false}
-            end
-
-          _ ->
-            {0, false, false}
-        end
-      end
-
-    cond do
-      not ok ->
-        {slot, kind_mismatch(report)}
-
-      true ->
-        report = if saturated, do: clamped(report), else: report
-        {value, report} = clamp_range(value, f, report)
-        clamp_storage(value, f, report)
-    end
-  end
-
-  defp saturate(d, f) do
-    signed = f.kind >= 2 and f.kind <= 5
-
-    cond do
-      not signed and d < 0.0 -> 0
-      not signed -> 0xFFFFFFFFFFFFFFFF
-      d < 0.0 -> -0x8000000000000000
-      true -> 0x7FFFFFFFFFFFFFFF
-    end
-  end
-
-  # converted digit by digit so no width and no locale can move it; saturation
-  # is reported as a clamp, the wire's rule for a value outside what the reader
-  # can hold (§4)
-  defp token_integer(token, f) do
-    signed = f.kind >= 2 and f.kind <= 5
-    {negative, digits} =
-      case token do
-        "-" <> rest -> {true, rest}
-        "+" <> rest -> {false, rest}
-        rest -> {false, rest}
-      end
-
-    magnitude = String.to_integer(digits)
-
-    cond do
-      not signed and negative -> {0, magnitude != 0, true}
-      not signed and magnitude > 0xFFFFFFFFFFFFFFFF -> {0xFFFFFFFFFFFFFFFF, true, true}
-      not signed -> {magnitude, false, true}
-      negative and magnitude > 0x8000000000000000 -> {-0x8000000000000000, true, true}
-      negative -> {-magnitude, false, true}
-      magnitude > 0x7FFFFFFFFFFFFFFF -> {0x7FFFFFFFFFFFFFFF, true, true}
-      true -> {magnitude, false, true}
-    end
-  end
-
-  # the field's own storage width is the last bound: a value past it clamps
-  # rather than wrapping, which is what the wire does too
-  defp clamp_storage(value, f, report) do
-    bytes = f.storage_bytes
-
-    if bytes <= 0 or bytes >= 8 do
-      {value, report}
-    else
-      signed = f.kind >= 2 and f.kind <= 5
-
-      if signed do
-        hi = (1 <<< (bytes * 8 - 1)) - 1
-        clamp(value, -hi - 1, hi, report)
-      else
-        clamp(value, 0, (1 <<< (bytes * 8)) - 1, report)
-      end
-    end
-  end
-
-  defp read_union(slot, f, text, pos, depth, value, report) do
-    # a union is an object with ONE key, the arm's name; {} is None, and two
-    # keys is a text this walk will not guess at
-    {c, pos} = peek(text, pos)
-    if c != ?{, do: bad(value, report)
-    arms = info(f.arms)
-    empty = info(f.union_new)
-    {c2, pos} = peek(text, pos + 1)
-
-    if c2 == ?} do
-      {empty, pos + 1, report}
-    else
-      {key, pos, _} = scan_string(text, pos, @max_key, value, report)
-      {colon, pos} = peek(text, pos)
-      if colon != ?:, do: bad(value, report)
-      pos = pos + 1
-
-      tag =
-        arms
-        |> Enum.with_index()
-        |> Enum.find_value(fn {arm, i} -> if i > 0 and arm.name == key, do: i, else: nil end)
-
-      {slot, pos, report} =
-        if tag == nil do
-          {slot, skip_value(text, pos, depth + 1, value, report), unknown(report)}
-        else
-          arm = Enum.at(arms, tag)
-          type = info(arm.table)
-          payload = info(type.new)
-          {payload, pos2, report2} = read_table(payload, type, text, pos, depth + 1, report)
-          {Map.put(Map.put(empty, :type, tag), arm.key, payload), pos2, report2}
-        end
-
-      {c3, pos} = peek(text, pos)
-      pos = if c3 == ?,, do: pos + 1, else: pos
-      {c4, pos} = peek(text, pos)
-      # a second key: a one-of with two arms is not a value
-      if c4 != ?}, do: bad(value, report)
-      {slot, pos + 1, report}
-    end
-  end
-
-  defp read_flags(f, text, pos, depth, value, report) do
-    {c, pos} = peek(text, pos)
-    if c != ?[, do: bad(value, report)
-    read_flag_names(f, text, pos + 1, depth, value, report, 0)
-  end
-
-  defp read_flag_names(f, text, pos, depth, value, report, bits) do
-    {c, pos} = peek(text, pos)
-
-    cond do
-      c == ?] ->
-        {bits, pos + 1, report}
-
-      c == 0 ->
-        bad(value, report)
-
-      true ->
-        {bits, pos, report} =
-          if c != ?" do
-            {bits, skip_value(text, pos, depth + 1, value, report), kind_mismatch(report)}
-          else
-            {name, pos2, _} = scan_string(text, pos, @max_key, value, report)
-
-            case Enum.find_index(f.flags, fn v -> v == name end) do
-              nil -> {bits, pos2, unknown(report)}
-              bit -> {bits ||| 1 <<< bit, pos2, report}
-            end
-          end
-
-        {c2, pos2} = peek(text, pos)
-
-        cond do
-          c2 == ?, -> read_flag_names(f, text, pos2 + 1, depth, value, report, bits)
-          c2 == ?] -> {bits, pos2 + 1, report}
-          true -> bad(value, report)
-        end
-    end
-  end
-`
